Switch from Float to Double when parsing Gerber data (#18)
We are losing some precision when parsing coordinate data from Gerber
and storing it in Float so we switch to Double to avoid this. A better
approach would probably have been a fixed precision representation but
Double gets rid of most of the precision loss while not significantly
changing the rest of the code.

Here is a hand wavy rough justification. `FSLAX45Y45` in MM means
coordinates have 4 integer + 5 decimal digits = 9 significant digits
max. The largest value is 9999.99999 mm.
- Double provides ~15.9 decimal digits of precision — more than enough for 9 digits. No loss.
- Float provides ~7.2 decimal digits — not enough for 9 digits. This was the bug.
- `Y14286369` decodes to 142.86369 mm (9 significant digits):
  - As Double: 142.86369 stored exactly
  - As Float: 142.86369 rounds to 142.8637 — the last 2 digits are lost

+++ gerber/lib/Gerber/ApertureDefinition.hs
@@ -10,24 +10,24 @@ data ApertureDefinition
 
 
 data CircleModifiers = CircleModifiers
-  { diameter :: !Float
-  , circleHoleDiameter :: !(Maybe Float)
+  { diameter :: !Double
+  , circleHoleDiameter :: !(Maybe Double)
   }
   deriving (Eq, Ord, Show)
 
 
 data RectangleModifiers = RectangleModifiers
-  { width :: !Float
-  , height :: !Float
-  , rectangleHoleDiameter :: !(Maybe Float)
+  { width :: !Double
+  , height :: !Double
+  , rectangleHoleDiameter :: !(Maybe Double)
   }
   deriving (Eq, Ord, Show)
 
 
 data PolygonModifiers = PolygonModifiers
-  { outerDiameter :: !Float
+  { outerDiameter :: !Double
   , numberOfVertices :: !Int
-  , rotation :: !(Maybe Float)
-  , polygonHoleDiameter :: !(Maybe Float)
+  , rotation :: !(Maybe Double)
+  , polygonHoleDiameter :: !(Maybe Double)
   }
   deriving (Eq, Ord, Show)
